🔨 refactor: traffics methods of traffic module now pass to common module as traffic
