+++ common/repository/traffic.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/metalpoch/olt-blueprint/common/entity"
-	"github.com/metalpoch/olt-blueprint/traffic/model"
+	"github.com/metalpoch/olt-blueprint/common/model"
 	"gorm.io/gorm"
 )
 
@@ -12,11 +12,15 @@ type trafficRepository struct {
 	db *gorm.DB
 }
 
-func NewUserRepository(db *gorm.DB) *trafficRepository {
+func NewTrafficRepository(db *gorm.DB) *trafficRepository {
 	return &trafficRepository{db}
 }
 
-func (repo trafficRepository) GetTrafficByInterface(ctx context.Context, id uint, date *model.RangeDate) ([]*entity.Traffic, error) {
+func (repo trafficRepository) Add(ctx context.Context, traffic *entity.Traffic) error {
+	return repo.db.WithContext(ctx).Create(traffic).Error
+}
+
+func (repo trafficRepository) GetTrafficByInterface(ctx context.Context, id uint, date *model.TranficRangeDate) ([]*entity.Traffic, error) {
 	var traffic []*entity.Traffic
 	err := repo.db.WithContext(ctx).
 		Where("interface_id = ? AND date BETWEEN ? AND ?", id, date.InitDate, date.EndDate).
@@ -26,7 +30,7 @@ func (repo trafficRepository) GetTrafficByInterface(ctx context.Context, id uint
 	return traffic, err
 }
 
-func (repo trafficRepository) GetTrafficByDevice(ctx context.Context, id uint, date *model.RangeDate) ([]*entity.Traffic, error) {
+func (repo trafficRepository) GetTrafficByDevice(ctx context.Context, id uint, date *model.TranficRangeDate) ([]*entity.Traffic, error) {
 	var traffic []*entity.Traffic
 	err := repo.db.WithContext(ctx).
 		Model(&entity.Traffic{}).
